Car rotates on the spot
After 2 days I have finally gotten the car to rotate on the spot. I
eventually realised that I could use the center positions from before
and after the rotation to see how much the car's image had moved during
the rotation. I had to do this because pygame will always rotate an
image around the top left hand corner. I could then use the difference
in x and y positions to find an offset, which I could then apply to the
car's x and y positions to move them back to where they were before the
rotation.
The reason I did this over just drawing the image to the screen in the
rotate function was to keep the simple drawing logic that I currently
use. This may also make things such as collision detection easier as the
x and y coordinates of the player will be correct.