Player now has friction
The player now has a friction force applied to their speed which slows
them down. Ideally this force would change from surface to surface e.g.
gravel has more friction than tarmac, but so far I don't know how to
implement this.

I would guess that the surface that the car is on would be passed in as
an argument to player.move() but this could cause issues if the wheels
are on different surfaces. I'm also unsure as how best to track where
the different surfaces are e.g. what part of the route is dirt vs grass.