Player car now moves in correct direction
The player car now moves in the direction it is facing. The issue I was
facing was that the car was rotated 90 degrees in the image, so it did
not line up with it's direction variable. I fixed this by simply
rotating the image 90 degrees.

I also had to use radians in the calculations, which was the final piece
of the puzzle.