Added finish banner and different surfaces
I added a finish banner which ends the game when the player collides
with it. However, the game does not currently restart at all after this
so that needs to be fixed.

The game now figures out which surface the player is on and then the
player has different grip and friction levels based on this. However,
the collision detection currently seems a bit strange, so I may have to
change it.

Lastly, I also did some more work on making the game scale well for
different screen sizes. However, the steering speed does not change with
the screen size and I am unsure if it should. Previously. it would scale
along with the screen size but this did not feel right. Currently, it
does not and this feels better but I am unsure if this accurate.